the end

+++ dars7/src/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, BadRequestException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -23,12 +23,16 @@ export class UsersController {
   }
 
   @Patch('/v2/update/:id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(@Param('id') id: string, @Body() updateUserDto: Partial<UpdateUserDto>) : ReturnType<UsersService["update"]> {
+    if(!updateUserDto || Object.values(updateUserDto).length === 0) throw new BadRequestException("Body is empty not allowwed !")
     return this.usersService.update(+id, updateUserDto);
   }
 
   @Delete('v3/delete/:id')
   remove(@Param('id') id: string) {
+    if(isNaN(+id)){
+      throw new BadRequestException("Invalid id !")
+    }
     return this.usersService.remove(+id);
   }
 }
